requirements/base.txt
# Base
pytz==2020.1
python-slugify==4.0.1
Pillow==7.2.0
psycopg2==2.8.5 --no-binary psycopg2

# Django
Django==3.1.1

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.11.1
django-filter==2.3.0

# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==5.2.0

# Celery
tornado>=5.0.0,<6.0.0
redis>=3.5.3
django-redis==4.12.1
celery==4.4.7
flower==0.9.5
